requirements.txt
pythoncertifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
requests==2.32.3
urllib3==2.2.3
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
python-decouple==3.8
requests==2.32.3
urllib3==2.2.3
